Renamed susebuild package to regionsrv
